[Breaking change] Implement sass caching
Adds a `cache_options` parameter to `sass()`, and a corresponding
`sass_cache_options` function. If running in a non-interactive R
session, or with `options(sass.cache=TRUE)`, or explicitly passing
cache options to `sass()`, then `sass()` will write and read cache
data to/from disk, to try to avoid actually compiling.

Breaking change: Previously, calling sass() with a non-NULL output
would result in the output file being written AND the CSS being
returned invisibly. With this change, the output file is written
but the CSS is no longer returned (just invisible()). This is to
avoid the expense of reading the CSS from disk if it is not likely
to be actually used.